* Replaced floats with doubles * Changed crit chance display to match ingame spellbook value (+0.04% per weaponskill over 300) * Formatted stats view values to display maximum of 2 digits

+++ src/main/java/sim/items/Enchant.java
@@ -59,8 +59,8 @@ public class Enchant implements Serializable {
         this.slot = slot;
     }
 
-    public float getHaste() {
-        return 1 + haste/100.0f;
+    public double getHaste() {
+        return 1 + haste/100.0d;
     }
 
     public void setHaste(int haste) {
